requirements.txt
django
django-heroku
djangorestframework
django-viewflow==2.0.0a2
gunicorn
pyjwt==1.5.3
whitenoise
dj-database-url
django-filter==0.9.2
django-filters==0.1.0